+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #363 Upgrade to JUnit 5.13.4
 * #364 "--config=stdin:xxx" - STDIN as an extra config source
+* #365 "BQCoreModuleExtender.setPropertyProvider(..)": deferred property creation
 
 ## 4.0-M1